// frontend/src/app/models/ucenik.ts
export class Ucenik{
    korisnicko_ime: string = "";
    lozinka: string = "";
    bezbedonosno_pitanje: string = "";
    odgovor: string = "";
    ime: string = "";
    prezime: string = "";
    pol: string = "";
    adresa: string = "";
    kontakt_telefon: string = "";
    email: string = "";
    profilna_slika: string = "";
    tip_skole: string = "";
    trenutni_razred: number = 0;
    ocene: Array<number> = [];
}